Sync IDC mitigation for multi-sites with mapped domains
Here is a summary of messages from 12 commits:

- Add get_raw_url method to Jetpack_Sync_Functions
- Add SSL support to Jetpack_Sync_Functions::get_raw_url()
- Allow using raw URL with constant
- Add filters to the home_url and site_url callables
- Add support for domain mapping
- 3rd Party: Hook domain mapping plugins to sync filters
- Add tests for 3rd party domain mapping support
- Fix broken test by changing constant to static variable
- Use Jetpack_Sync_Options to bypass potential caching issues
- Remove error_log that was causing tests to fail
- Swap DOMAIN_MAPPING constant check to SUNRISE_LOADED
- Default to using raw home and siteurl values

Notes:
In the WordPress MU Domain Mapping plugin, the DOMAIN_MAPPING constant is only set if we're actively mapping a domain. So, if thisdomain.com points to somesubsite.network.com. This led to the 3rd party domain mapping logic to fail for site where the admin was not using the mapped domain.

The SUNRISE_LOADED constant is loaded on every request with the WordPress MU Domain Mapping plugin.